menambahkan subclass baru : EmployeeJoined, EmployeeSalary, EmployeeSpouse
